requirements.txt
customtkinter
yfinance
pandas
pandas_ta
google-genai
python-dotenv
packaging
cerebras_cloud_sdk
groq
openai
matplotlib
mplfinance
tkinterweb
